add back button and niceedit

+++ modules/authorization/edit.php
@@ -1,5 +1,12 @@
 <?php
 
+//if not admin
+if($_SESSION['ROLE']>2){
+    die($tpl->render("alert",['msg'=>'Unauthorized','msgType'=>'warning',
+                    'url'=>'/'.$crumbs[0].'/'.$crumbs[1]."/".$id]));
+}
+
+
 if(isset($_POST['delete']) && $_POST['delete']==true){
     if($_db->delete('api_auth',['id'=>$_path[0]])->rowCount()>0){
         die($tpl->render("alert",['msg'=>'Auth Deleted','msgType'=>'success','url'=>'/'.$crumbs[0].'/'.$crumbs[1].'/']));
